requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
ecdsa==0.18.0
fastapi==0.104.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytest==7.4.2
pytest-order==1.1.0
pytest-ordering==0.6
python-dotenv==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
